requirements.txt
flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.26
werkzeug==2.0.3
matplotlib==3.5.2 